Use same styling for all editor pages
