--- src/components/BackButton/styles.scss
@@ -1,16 +1,0 @@
-.back-button {
-  button {
-    border: none;
-    display: flex;
-    align-items: baseline;
-    min-width: fit-content;
-
-    svg {
-      margin-right: 1rem;
-    }
-  }
-
-  button:hover {
-    text-decoration: underline;
-  }
-}
