--- src/components/PageHeader/styles.scss
@@ -1,3 +1,0 @@
-.page-header {
-  display: flex;
-}
