+++ src/scenes/Lines/scenes/Editor/styles.scss
@@ -1,16 +1,6 @@
 @import '~@entur/tokens/dist/styles.scss';
 
-.editor-frame,
 .line-editor {
-  margin: 0 5rem;
-}
-
-.editor-frame {
-  margin-top: 2rem;
-}
-
-.line-editor {
-  margin-top: 2rem;
   flex: 1;
   display: flex;
   flex-direction: column;
